Fix location of project and survey delete buttons (#379)
